+++ app/assets/config/universidade_manifest.js
@@ -1,2 +1,3 @@
 //= link_directory ../stylesheets/universidade .css
 //= link_tree ../images/universidade .svg
+//= link universidade/application.js